--- src/leaderboard/shared/constants.ts
@@ -1,1 +1,0 @@
-export const MIN_PUBLIC_SCORE = 1.50
